% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_template.R
\name{add_template}
\alias{add_template}
\title{Copies Template for User Research Process}
\usage{
add_template(rdir = "./R")
}
\arguments{
\item{rdir}{the directory to copy the template to}
}
\description{
Copies Template for User Research Process.
}
\examples{
\dontrun{
add_template()
}
}

