% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_process.R
\name{batch_process}
\alias{batch_process}
\title{Creates Code to Activate Multiple Functions and Copies to Clipboard}
\usage{
batch_process(url, ...)
}
\arguments{
\item{url}{the url of the google form to be pinged}

\item{\ldots}{the function(s) to be activated}
}
\description{
Creates Code to Activate Multiple Functions and Copies to Clipboard.
}
\examples{
batch_process(url = "dummy_url", "fun1", "fun2")
# the output is already copied to clipboard on most systems
}

