% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pingr.R
\name{pingr}
\alias{pingr}
\title{Activates Any Function with User Research Capability}
\usage{
pingr(fun)
}
\arguments{
\item{fun}{the function to be activated}
}
\description{
Activates Any Function with User Research Capability.
}

